Easy ~1.2x speedup for calculateRadius
This adds an optimization to avoid calculating where you can go from a position if you have already maxed out. I am not aware of the cost structure, but if there is a positive, minimum cost to making a step (as in, if each moved required at least 1), then you could add that minimum to the `nextCost` in the if statement.

There is no formal benchmarking system in the repo, so this just used a homebrewed one. The (fairly unscientific) results were speedups ranging from 1.15x to 1.3x